First commit
For the record, i made this years ago and just adapted it to work using PySide6.4.
There are places to generate NPCs (for the story or ennemies to battle) with unique names, vocation and stats (The stats of the NPCs are identical but useful for battle since you can roll a dice and having their health change accordingly).

I'm planning to make a modified version of this for iOS.